ref: split func
